refactor: extract app options
