+++ test/util/testnode/config.go
@@ -6,8 +6,6 @@ import (
 	"github.com/celestiaorg/celestia-app/cmd/celestia-appd/cmd"
 	"github.com/celestiaorg/celestia-app/pkg/appconsts"
 	"github.com/celestiaorg/celestia-app/test/util/genesis"
-	pruningtypes "github.com/cosmos/cosmos-sdk/pruning/types"
-	"github.com/cosmos/cosmos-sdk/server"
 	srvconfig "github.com/cosmos/cosmos-sdk/server/config"
 	srvtypes "github.com/cosmos/cosmos-sdk/server/types"
 	tmconfig "github.com/tendermint/tendermint/config"
@@ -127,27 +125,6 @@ func DefaultConfig() *Config {
 		WithAppOptions(DefaultAppOptions()).
 		WithAppCreator(cmd.NewAppServer).
 		WithSupressLogs(true)
-}
-
-type KVAppOptions struct {
-	options map[string]interface{}
-}
-
-// Get returns the option value for the given option key.
-func (ao *KVAppOptions) Get(option string) interface{} {
-	return ao.options[option]
-}
-
-// Set sets a key-value app option.
-func (ao *KVAppOptions) Set(option string, value interface{}) {
-	ao.options[option] = value
-}
-
-// DefaultAppOptions returns the default application options.
-func DefaultAppOptions() *KVAppOptions {
-	opts := &KVAppOptions{options: make(map[string]interface{})}
-	opts.Set(server.FlagPruning, pruningtypes.PruningOptionNothing)
-	return opts
 }
 
 func DefaultParams() *tmproto.ConsensusParams {
